+++ pyproject.toml
@@ -8,7 +8,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "automakemkv"
-version = "3.4.0"
+version = "3.5.0"
 description = "Package for automagically ripping media with MakeMKV"
 readme = "README.md"
 authors = [
